Add SES network to default networks (SOC-10451)
This commit adds an ses_client network to the network
barclamp's default data bag. This causes that particular
network to be created for any new deployment. Clouds with SES
integration enabled, the ses_client network will be
automatically connected to any nodes with roles that will need
access to the SUSE Enterprise Storage (glance-server or
cinder-volume).