upgrade: cold-start Nova for normal mode
In the normal upgrade mode, Nova will have to be cold started
just the same as in non-disruptive mode, i.e. after the Nova
services have been updated on both the controllers _and_ the
compute nodes. Otherwise, there will be records of offline
compute nodes with the old RPC API version at the time of
controller service restart, causing the controller services to
use 4.17 and enforce that same version for compute nodes
later.